requirements.txt
annotated-types==0.6.0
anyio==3.7.1
azure-applicationinsights==0.1.1
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-search-documents==11.3.0
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
certifi==2023.7.22
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
exceptiongroup==1.2.0
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
openai==0.27.4
openapi-schema-pydantic==1.2.4
packaging==23.2
pluggy==1.3.0
priority==2.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyMySQL==1.1.0
pyodbc==5.1.0
pytest==7.4.2
python-decouple==3.8
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
tomli==2.0.1
typing==3.7.4.3
typing_extensions==4.8.0
urllib3==2.2.1
uvicorn==0.23.2
wsproto==1.2.0
